producer/requirements.txt
google-cloud-pubsub==2.18.0
google-cloud-bigquery==3.11.0
google-cloud-storage==2.10.0
requests==2.31.0
python-dotenv==1.0.0
